+++ common/src/test/java/io/netty/util/DomainNameMappingTest.java
@@ -107,22 +107,22 @@ public class DomainNameMappingTest {
 
     @Test(expected = NullPointerException.class)
     public void testNullDefaultValue() {
-        new DomainMappingBuilder<String>(null);
+        new DomainNameMappingBuilder<String>(null);
     }
 
     @Test(expected = NullPointerException.class)
     public void testNullDomainNamePatternsAreForbidden() {
-        new DomainMappingBuilder<String>("NotFound").add(null, "Some value");
+        new DomainNameMappingBuilder<String>("NotFound").add(null, "Some value");
     }
 
     @Test(expected = NullPointerException.class)
     public void testNullValuesAreForbidden() {
-        new DomainMappingBuilder<String>("NotFound").add("Some key", null);
+        new DomainNameMappingBuilder<String>("NotFound").add("Some key", null);
     }
 
     @Test
     public void testDefaultValue() {
-        DomainNameMapping<String> mapping = new DomainMappingBuilder<String>("NotFound")
+        DomainNameMapping<String> mapping = new DomainNameMappingBuilder<String>("NotFound")
             .add("*.netty.io", "Netty")
             .build();
 
@@ -131,7 +131,7 @@ public class DomainNameMappingTest {
 
     @Test
     public void testStrictEquality() {
-        DomainNameMapping<String> mapping = new DomainMappingBuilder<String>("NotFound")
+        DomainNameMapping<String> mapping = new DomainNameMappingBuilder<String>("NotFound")
             .add("netty.io", "Netty")
             .add("downloads.netty.io", "Netty-Downloads")
             .build();
@@ -144,7 +144,7 @@ public class DomainNameMappingTest {
 
     @Test
     public void testWildcardMatchesAnyPrefix() {
-        DomainNameMapping<String> mapping = new DomainMappingBuilder<String>("NotFound")
+        DomainNameMapping<String> mapping = new DomainNameMappingBuilder<String>("NotFound")
             .add("*.netty.io", "Netty")
             .build();
 
@@ -158,14 +158,14 @@ public class DomainNameMappingTest {
     @Test
     public void testFirstMatchWins() {
         assertEquals("Netty",
-            new DomainMappingBuilder<String>("NotFound")
+            new DomainNameMappingBuilder<String>("NotFound")
                 .add("*.netty.io", "Netty")
                 .add("downloads.netty.io", "Netty-Downloads")
                 .build()
                 .map("downloads.netty.io"));
 
         assertEquals("Netty-Downloads",
-            new DomainMappingBuilder<String>("NotFound")
+            new DomainNameMappingBuilder<String>("NotFound")
                 .add("downloads.netty.io", "Netty-Downloads")
                 .add("*.netty.io", "Netty")
                 .build()
@@ -174,7 +174,7 @@ public class DomainNameMappingTest {
 
     @Test
     public void testToString() {
-        DomainNameMapping<String> mapping = new DomainMappingBuilder<String>("NotFound")
+        DomainNameMapping<String> mapping = new DomainNameMappingBuilder<String>("NotFound")
             .add("*.netty.io", "Netty")
             .add("downloads.netty.io", "Netty-Download")
             .build();
@@ -199,7 +199,7 @@ public class DomainNameMappingTest {
 
     @Test
     public void testAsMapWithImmutableDomainNameMapping() {
-        DomainNameMapping<String> mapping = new DomainMappingBuilder<String>("NotFound")
+        DomainNameMapping<String> mapping = new DomainNameMappingBuilder<String>("NotFound")
             .add("netty.io", "Netty")
             .add("downloads.netty.io", "Netty-Downloads")
             .build();
